Chore: combine tests in tests/lib/eslint.js and tests/lib/linter.js
3ec436eeed0b0271e2ed0d0cb22e4246eb15f137 renamed lib/eslint.js to lib/linter.js, but it did not rename the corresponding test file, which remained as tests/lib/eslint.js. When adding tests for linter.js afterwards, we put them in a new tests/lib/linter.js file. This is confusing because the tests for lib/linter.js are now in two different places. This commit combines the two test files into tests/lib/linter.js.
